+++ components/nav/Navbar.tsx
@@ -1,24 +1,18 @@
 'use client'
 
-import SidenavPage from '@/components/nav/SideNav';
+import SideNavHeader from '@/components/nav/sideNav/header';
 import useMediaQuery from '@/components/listeners/WidthSettings';
 import { usePathname } from 'next/navigation';
-import SidenavMobile from '@/components/nav/SideNavMobile';
 
 
 const NavBar = () => {
     const pathName = usePathname();
-    const isBreakpoint = useMediaQuery(768);
+
     return (
         <div className={`flex flex-row h-24 items-center w-full px-10`} style={{height: '8vh'}}>
-          {!isBreakpoint && pathName !== 'demo' &&
-          <SidenavPage/>
+          {pathName !== 'demo' &&
+          <SideNavHeader/>
           }
-          
-          {isBreakpoint && pathName !== 'demo' &&
-          <SidenavMobile/>
-          }
-
         </div>
     )
 }
